backend/requirements.txt
google-genai
google.adk
pymongo
python-dotenv
fastapi
uvicorn[standard]
certifi
redis
websockets
